refactor errorReducer to its own file
